axe all prev usenostr refs - new nostr class etc
