+++ packages/app/features/user/CreateAccountScreen.tsx
@@ -28,16 +28,6 @@ export const CreateAccountScreen = () => {
     signup(username, displayName, about)
   }, [username, displayName, about])
 
-  //   const { connect } = useNostr()
-
-  //   const connectem = async () => {
-  //     connect(['wss://relay.nostr.ch', 'wss://arc1.arcadelabs.co'])
-  //   }
-
-  //   useEffect(() => {
-  //     connectem()
-  //   }, [])
-
   return (
     <Screen>
       <BackButton mt={40} ml={20} />
