+++ packages/app/lib/nostr/index.ts
@@ -1,0 +1,2 @@
+export * from './Event'
+export * from './Nostr'
