+++ packages/app/package.json
@@ -23,6 +23,7 @@
     "expo-clipboard": "^4.0.1",
     "expo-linking": "^3.2.3",
     "expo-secure-store": "^12.0.0",
+    "nostr-relaypool": "0.2.1",
     "react-use-clipboard": "1.0.9",
     "solito": "2.0.0-canary.4"
   }
