+++ packages/app/stores/index.ts
@@ -1,3 +1,4 @@
+import { Nostr } from 'app/lib'
 import create from 'zustand'
 import { createAuthStore } from './auth'
 import { createChatStore } from './chat'
@@ -6,6 +7,7 @@ import { createRelayStore } from './relay'
 import { createUiStore } from './ui'
 
 export const useStore = create((set, get) => ({
+  nostr: new Nostr(),
   ...createAuthStore(set, get),
   ...createChatStore(set, get),
   ...createDatabaseStore(set, get),
